some long to int warnings from HP-UX log fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_2_BRANCH@7068 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/fs_mem.cpp
@@ -159,7 +159,7 @@ bool wxMemoryFSHandler::CheckHash(const wxString& filename)
 
     
     wxMemoryOutputStream mems;
-    if (image.Ok() && image.SaveFile(mems, type))
+    if (image.Ok() && image.SaveFile(mems, (int)type))
         m_Hash -> Put(filename, new MemFSHashObj(mems));
     else
     {
